Move metadata into its own file
Move `Metadata`, `MetadataLoweringError` and `ArchiveMetadata` into their own file `metadata.rs` in `uv-distribution`, moving it out from `lib.rs`. No functional changes.
